backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==1.4.52
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9